# swagger-doc/swagger.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/sample":{"post":{"tags":["my-controller"],"summary":"create sample object","description":"API to create sample object","operationId":"create sample object","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SampleObj"}}},"required":true},"responses":{"201":{"description":"post sample object successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SampleObj"}}}},"400":{"description":"Unsupported JSON format","content":{"application/json":{"schema":{"type":"object"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/sample/{id}":{"get":{"tags":["my-controller"],"summary":"Get sample object","description":"API to Get sample object","operationId":"Get sample object","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"201":{"description":"Get sample object successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SampleObj"}}}},"400":{"description":"Unsupported JSON format","content":{"application/json":{"schema":{"type":"object"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"SampleObj":{"type":"object","properties":{"name":{"type":"string"},"age":{"type":"string"},"address":{"type":"string"}}},"ContentDisposition":{"type":"object","properties":{"type":{"type":"string"},"name":{"type":"string"},"filename":{"type":"string"},"charset":{"type":"string"},"size":{"type":"integer","format":"int64","deprecated":true},"creationDate":{"type":"string","format":"date-time","deprecated":true},"modificationDate":{"type":"string","format":"date-time","deprecated":true},"readDate":{"type":"string","format":"date-time","deprecated":true},"inline":{"type":"boolean"},"attachment":{"type":"boolean"},"formData":{"type":"boolean"}}},"ErrorResponse":{"type":"object","properties":{"headers":{"type":"object","properties":{"origin":{"type":"string"},"contentType":{"$ref":"#/components/schemas/MediaType"},"connection":{"type":"array","items":{"type":"string"}},"empty":{"type":"boolean"},"location":{"type":"string","format":"uri"},"host":{"type":"object","properties":{"address":{"type":"object","properties":{"hostAddress":{"type":"string"},"address":{"type":"string","format":"byte"},"hostName":{"type":"string"},"linkLocalAddress":{"type":"boolean"},"multicastAddress":{"type":"boolean"},"anyLocalAddress":{"type":"boolean"},"loopbackAddress":{"type":"boolean"},"siteLocalAddress":{"type":"boolean"},"mcglobal":{"type":"boolean"},"mcnodeLocal":{"type":"boolean"},"mclinkLocal":{"type":"boolean"},"mcsiteLocal":{"type":"boolean"},"mcorgLocal":{"type":"boolean"},"canonicalHostName":{"type":"string"}}},"port":{"type":"integer","format":"int32"},"unresolved":{"type":"boolean"},"hostName":{"type":"string"},"hostString":{"type":"string"}}},"all":{"type":"object","additionalProperties":{"type":"string"},"writeOnly":true},"lastModified":{"type":"integer","format":"int64"},"contentLength":{"type":"integer","format":"int64"},"date":{"type":"integer","format":"int64"},"ifModifiedSince":{"type":"integer","format":"int64"},"range":{"type":"array","items":{"$ref":"#/components/schemas/HttpRange"}},"allow":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/HttpMethod"}},"cacheControl":{"type":"string"},"acceptCharset":{"type":"array","items":{"type":"string"}},"contentDisposition":{"$ref":"#/components/schemas/ContentDisposition"},"acceptPatch":{"type":"array","items":{"$ref":"#/components/schemas/MediaType"}},"acceptLanguage":{"type":"array","items":{"type":"object","properties":{"range":{"type":"string"},"weight":{"type":"number","format":"double"}}}},"basicAuth":{"type":"string","writeOnly":true},"accept":{"type":"array","items":{"$ref":"#/components/schemas/MediaType"}},"contentLanguage":{"type":"object","properties":{"language":{"type":"string"},"script":{"type":"string"},"country":{"type":"string"},"variant":{"type":"string"},"unicodeLocaleAttributes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"unicodeLocaleKeys":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"displayLanguage":{"type":"string"},"displayScript":{"type":"string"},"displayCountry":{"type":"string"},"displayVariant":{"type":"string"},"displayName":{"type":"string"},"extensionKeys":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"iso3Language":{"type":"string"},"iso3Country":{"type":"string"}}},"expires":{"type":"integer","format":"int64"},"ifMatch":{"type":"array","items":{"type":"string"}},"etag":{"type":"string"},"ifNoneMatch":{"type":"array","items":{"type":"string"}},"bearerAuth":{"type":"string","writeOnly":true},"vary":{"type":"array","items":{"type":"string"}},"pragma":{"type":"string"},"upgrade":{"type":"string"},"accessControlRequestHeaders":{"type":"array","items":{"type":"string"}},"accessControlExposeHeaders":{"type":"array","items":{"type":"string"}},"accessControlAllowCredentials":{"type":"boolean"},"accessControlAllowMethods":{"type":"array","items":{"$ref":"#/components/schemas/HttpMethod"}},"accessControlAllowHeaders":{"type":"array","items":{"type":"string"}},"accessControlMaxAge":{"type":"integer","format":"int64"},"accessControlAllowOrigin":{"type":"string"},"accessControlRequestMethod":{"$ref":"#/components/schemas/HttpMethod"},"ifUnmodifiedSince":{"type":"integer","format":"int64"},"acceptLanguageAsLocales":{"type":"array","items":{"type":"object","properties":{"language":{"type":"string"},"script":{"type":"string"},"country":{"type":"string"},"variant":{"type":"string"},"unicodeLocaleAttributes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"unicodeLocaleKeys":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"displayLanguage":{"type":"string"},"displayScript":{"type":"string"},"displayCountry":{"type":"string"},"displayVariant":{"type":"string"},"displayName":{"type":"string"},"extensionKeys":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"iso3Language":{"type":"string"},"iso3Country":{"type":"string"}}}}},"additionalProperties":{"type":"array","items":{"type":"string"}}},"body":{"$ref":"#/components/schemas/ProblemDetail"},"statusCode":{"$ref":"#/components/schemas/HttpStatusCode"},"titleMessageCode":{"type":"string"},"detailMessageArguments":{"type":"array","items":{"type":"object"}},"detailMessageCode":{"type":"string"},"typeMessageCode":{"type":"string"}}},"HttpMethod":{"type":"object"},"HttpRange":{"type":"object"},"HttpStatusCode":{"type":"object","properties":{"error":{"type":"boolean"},"is4xxClientError":{"type":"boolean"},"is5xxServerError":{"type":"boolean"},"is2xxSuccessful":{"type":"boolean"},"is3xxRedirection":{"type":"boolean"},"is1xxInformational":{"type":"boolean"}}},"MediaType":{"type":"object","properties":{"type":{"type":"string"},"subtype":{"type":"string"},"parameters":{"type":"object","additionalProperties":{"type":"string"}},"qualityValue":{"type":"number","format":"double"},"charset":{"type":"string"},"wildcardSubtype":{"type":"boolean"},"wildcardType":{"type":"boolean"},"concrete":{"type":"boolean"},"subtypeSuffix":{"type":"string"}}},"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{"type":"object"}}}}}}}